requirements.txt
PyGithub==2.1.1
GitPython==3.1.40